backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
fastapi-users[sqlalchemy]==14.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pyyaml==6.0.2
reportlab==4.4.0
google-api-python-client==2.170.0
google-auth==2.40.1
python-multipart==0.0.20
